Clone SIS2 in a separate state
Move the SIS2 repo clone step into a separate stage, which happens
before the build rules.

Previously, each build would attempt to re-clone the repository, which
created potential IO contention.  It also slowed each build step,
potentially past the 15 minute maximum.

This patch closely follows the MOM6 CI and clones the test suite to a
common directory set in JOB_DIR.  The SIS2 repo points to the local
version in CI_PROJECT_DIR, which contains the version to be tested.