+Rename checksum scale argument to unscale
  Renamed the scale argument to the various checksum calls used by SIS2 to
unscale, and added unscale optional arguments to the various check_redundant
routines.  Both of these changes bring SIS2 into line with what is used now in
MOM6 to facilitate the tighter integration of SIS2 and MOM6.  All calls to
checksums had the name of the scale optional argument changed, and several calls
to the check_redundant routines now have unscale arguments.  In addition,
descriptions of the units of about 55 real SIS2 variables were added, mostly by
copying comments directly from the corresponding MOM6 routines.  All answers are
bitwise identical, but there is a change in the name of an optional argument to
a number of publicly visible routines.