+Add set_segnum_signs and OBC loop ranges on PE
  Added 16 integer elements to the ocean_OBC_type to indicate the ranges of
indices over which various directions of open boundary conditions are found in
the data and computational domain of the local PE. Also added 4 logical elements
that indicate whether each of the 4 orientations of OBCs are fond on the local
PE.  These are now all being set, along with the direction-indicating signs of
segnum_u and segnum_v in the new subroutine set_segnum_signs(), which is a near
copy of the MOM_open_boundary.F90 version. All answers are bitwise identical,
but there are 16 new integer elements and 4 new logical elements in a
transparent public type.