*Fix negative zero maximum CFL
  SIS2 reports the global maximum CFL in the seaice.stats files, but in the
first timestep when all velocities are zero, it is reporting a maximum CFL of
-0.000 with the gnu and Nvidia compilers.  This single-line commit replaces any
negative zero maximum CFL returned by `max_across_PEs()` with its absolute value
so that only non-negative maximum CFL values are returned.  All solutions are
bitwise identical, but this does change the entries in the seaice.stats files so
it may appear to fail automated regression testing in some cases.