+Change some SIS2 axis units to CMIP standards
  Modified set_SIS_axes_info to follow the pattern from MOM6 and take the units
for the horizontal axes from the grid type.  In most cases, this has the effect
of changing the units from "degrees_E" and "degrees_N" to "degrees_east" and
"degrees_north" to follow CMIP protocols, and to mirror changes that were made
to MOM6 in 2017.  All solutions are bitwise identical, but there are changes to
the metadata in some output files, and some runtime parameters that had been
read from both `set_SIS_axes_info()` and `set_grid_metrics()` in
MOM6/src/initialization/MOM_grid_initialize.F90 are now only read in the latter
routine.  As a result of these changes, `set_SIS_axes_info()` no longer takes
a `param_file_type` argument and the grid arguments are now `intent(in)`.