Fix unit scaling in SIS_sponge
  Corrected the unit scaling factor for reading the sponge relaxation rate to
US%T_to_s as is appropriate for a rate.  An incorrect but consistent
compensating factor was removed from the expression for max_rlxrate, and the
line unscaling dt_slow into a new variable was also removed.

  There are some other minor changes included in this commit.  The unused
variables I_Nk, coeff, Idt_slow, ice_salin, iconc_old, ithk_old, mask_in and 6
others were removed from apply_isponge and a three-line expression setting the
limited fractional open water category was consolidated down to a single line.
The missing rescaling factor for the initial value of tfi was added and the
initial value of sice was change to 0; neither of these initial values appear to
be used.  In addition comments describing the units of 23 real variables were
added in SIS_sponge.F90.

  All answers are bitwise identical when no dimensional rescaling is applied,
but more of the code is now working with rescaled variables.